# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PadelPro API
  version: 0.1.0
  description: |
    API REST para el sistema de gestión de reservas de pádel PadelPro.
    Sistema On-Premise para instalaciones pequeñas (v1.0).

    ### Autenticación
    Se usa JWT HS256 con doble token (access + refresh). El access token dura 15 minutos.
    El refresh token dura 7 días y se almacena en cookie httpOnly.

    ### RBAC
    - `ADMIN`: acceso total, incluidos endpoints bajo `/api/admin/`.
    - `USER`: acceso a sus propios recursos (reservas, pagos, perfil).
    - Endpoints públicos: no requieren autenticación (marcados con `security: []`).
  contact:
    name: orquestadoria
    url: https://github.com/lcasadov/PadelPro
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api
    description: "Entorno de desarrollo local"
  - url: https://pre.padelpro.local/api
    description: "Entorno de preproducción"
  - url: https://padelpro.local/api
    description: "Entorno de producción"
security:
  - bearerAuth: []
...
paths:
...
  /reservas:
    get:
      tags: [Reservas]
      operationId: listarMisReservas
      summary: Listar reservas del usuario autenticado
      description: |
        Lista las reservas en las que el usuario autenticado es propietario o participante,
        con paginación. Requiere ROLE_USER o ROLE_ADMIN.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
      responses:
        '200':
          description: Lista devuelta correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReservasResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Reservas]
      operationId: crearReserva
      summary: Crear reserva
      description: |
        Crea una nueva reserva. El usuario autenticado es automáticamente añadido
        como primer participante (titular). Se envía OTP de tipo RESERVATION_CONFIRM
        al creador vía Telegram si la reserva proviene del canal Telegram.
        Requiere ROLE_USER o ROLE_ADMIN.

        **Idempotencia (RN-RES-05):** envíe la cabecera `Idempotency-Key`. Un segundo `POST`
        con la misma clave y el mismo usuario devuelve la reserva ya creada (201) sin duplicarla.

        **Precio (RN-RES-03):** el importe total se calcula exclusivamente en el backend
        (`price_per_hour × durationMinutes / 60`) y se congela en el pago asociado; cualquier
        importe enviado por el cliente se ignora.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: |
            Clave de idempotencia por usuario (RN-RES-05). Reintentos con la misma clave
            devuelven la misma reserva sin crear duplicados.
          schema:
            type: string
            maxLength: 255
            example: "abc-123-e89b-12d3"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservaRequest'
            examples:
              ejemplo:
                summary: Reserva de 60 min con un participante adicional externo
                value:
                  reservationDate: "2025-06-15"
                  startTime: "09:00"
                  durationMinutes: 60
                  notes: "Reserva para partido de entrenamiento"
                  participantesAdicionales:
                    - externalName: "Carlos García"
      responses:
        '201':
          description: Reserva creada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservaResponse'
              examples:
                ejemplo:
                  summary: Reserva creada en estado PENDING_CONFIRMATION
                  value:
                    id: "550e8400-e29b-41d4-a716-446655440000"
                    reservationDate: "2025-06-15"
                    startTime: "09:00"
                    durationMinutes: 60
                    status: "PENDING_CONFIRMATION"
                    ownerId: 5
                    ownerName: "John Doe"
                    priceTotal: 15.00
                    notes: "Reserva para partido de entrenamiento"
                    participants:
                      - id: 1
                        userId: 5
                        nombre: "John Doe"
                        statusPago: "PENDING"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: |
            La franja horaria solapa con una reserva activa (RN-RES-01). El constraint
            `excl_res_no_overlap` (gist) garantiza la exclusión incluso bajo concurrencia.
            Cuerpo: `code: "CONFLICT"`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            Regla de negocio incumplida. Códigos posibles:
            `PARTICIPANTS_LIMIT_EXCEEDED` (excede `max_participants_per_pista`),
            `INVALID_STATE_TRANSITION`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservas/{id}:
    get:
      tags: [Reservas]
      operationId: getReserva
      summary: Obtener detalle de reserva
      description: |
        Devuelve el detalle de una reserva. Un USER solo puede acceder si es propietario
        o participante. ADMIN puede acceder a cualquier reserva.
      parameters:
        - $ref: '#/components/parameters/ReservationIdParam'
      responses:
        '200':
          description: Reserva encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservaResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Reservas]
      operationId: cancelarReserva
      summary: Cancelar reserva
      description: |
        Cancela una reserva. Solo puede hacerlo el propietario (o ADMIN).
        La cancelación debe realizarse con al menos `cancellation_deadline_hours` de antelación.
        Se envía OTP de tipo CANCELLATION_CONFIRM al propietario vía Telegram.
      parameters:
        - $ref: '#/components/parameters/ReservationIdParam'
      responses:
        '204':
          description: Reserva cancelada correctamente
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: La cancelación está fuera del plazo permitido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservas/{id}/unirse:
    post:
      tags: [Reservas]
      operationId: unirseAReserva
      summary: Unirse a una reserva como participante
      description: |
        Permite a un usuario autenticado unirse a una reserva existente que tiene huecos disponibles.
        Un usuario no puede unirse dos veces a la misma reserva.
        La reserva debe estar en estado PENDING_CONFIRMATION o CONFIRMED.
        Requiere ROLE_USER o ROLE_ADMIN.
      parameters:
        - $ref: '#/components/parameters/ReservationIdParam'
      responses:
        '200':
          description: Usuario añadido como participante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnirseResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: |
            El usuario ya participa en la reserva (RN-AUTH-03). Error code `CONFLICT`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: |
            La reserva está completa (`PARTICIPANTS_LIMIT_EXCEEDED`, RN-RES-02) o su estado no admite
            unión (`RESERVA_NOT_JOINABLE`; CANCELLED/COMPLETED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reservas/{id}/participacion:
    delete:
      tags: [Reservas]
      operationId: abandonarReserva
      summary: Abandonar una partida (participante no-owner)
      description: |
        Permite a un participante no-owner abandonar una reserva a la que se unió, liberando su plaza.
        El titular (owner) NO puede abandonar: debe cancelar la reserva mediante el flujo de
        cancelación. Requiere ROLE_USER o ROLE_ADMIN.
      parameters:
        - $ref: '#/components/parameters/ReservationIdParam'
      responses:
        '204':
          description: El participante ha abandonado la reserva (plaza liberada)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          description: |
            El titular no puede abandonar (`OWNER_CANNOT_ABANDON`), el usuario no participa en la
            reserva (`NOT_A_PARTICIPANT`) o la reserva no está activa (`RESERVA_NOT_JOINABLE`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ─── Reservas — administración ─────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT HS256 obtenido en POST /auth/login o POST /auth/refresh.
        Incluir en la cabecera: `Authorization: Bearer <token>`.
        Lifetime: 15 minutos.

    redsysWebhookSecret:
      type: apiKey
      in: header
      name: X-Redsys-Signature
      description: Firma HMAC SHA-256 de Redsys para validar el origen del webhook de notificación de pago.

    telegramWebhookSecret:
      type: apiKey
      in: header
      name: X-Telegram-Bot-Api-Secret-Token
      description: Secret token configurado al registrar el webhook del bot de Telegram.

  # ─── Reusable Parameters ──────────────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      description: Índice de la página (base 0)
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
    PageSizeParam:
      name: pageSize
      in: query
      description: Número de elementos por página (máximo 100)
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ReservationIdParam:
      name: id
      in: path
      description: UUID de la reserva
      required: true
      schema:
        type: string
        format: uuid
        example: "550e8400-e29b-41d4-a716-446655440000"
...
  responses:
    Unauthorized:
      description: Token ausente, expirado o inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            tokenExpirado:
              summary: Token expirado
              value:
                code: "UNAUTHORIZED"
                message: "El token de acceso ha expirado."
                errors: []
    Forbidden:
      description: El usuario autenticado no tiene permisos para esta operación
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            sinRol:
              summary: Rol insuficiente
              value:
                code: "FORBIDDEN"
                message: "No tiene permisos para realizar esta operación."
                errors: []
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            noEncontrado:
              summary: Recurso no encontrado
              value:
                code: "NOT_FOUND"
                message: "El recurso solicitado no existe."
                errors: []
    ValidationError:
      description: Error de validación de parámetros o cuerpo de la petición
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            campoInvalido:
              summary: Campo con formato inválido
              value:
                code: "VALIDATION_ERROR"
                message: "Los datos de la solicitud no son válidos."
                errors:
                  - field: "email"
                    message: "El formato del email no es válido."
...
  schemas:
...
    ReservationStatus:
      type: string
      description: Estado del ciclo de vida de una reserva
      enum:
        - PENDING_CONFIRMATION
        - CONFIRMED
        - CANCELLED
        - COMPLETED
      example: CONFIRMED
    ReservationChannel:
      type: string
      description: Canal por el que se creó o se unió a la reserva
      enum:
        - WEB
        - TELEGRAM
      example: WEB
    PaymentStatus:
      type: string
      description: Estado del ciclo de pago
      enum:
        - PENDING
        - IN_PROGRESS
        - PAID
        - FAILED
        - CANCELLED
        - REFUNDED
      example: PAID
    PaymentMethod:
      type: string
      description: Método de pago utilizado
      enum:
        - REDSYS
        - CASH
      example: REDSYS
...
    ErrorResponse:
      type: object
      description: >-
        Respuesta de error estándar de la API, emitida por GlobalExceptionHandler.
        El código de aplicación va en `error`; `details` sólo aparece en errores que
        aportan una lista de violaciones (p. ej. INVALID_PASSWORD) y se omite en el
        resto (JSON con `NON_NULL`).
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Código de error de aplicación
          example: VALIDATION_ERROR
        message:
          type: string
          description: Descripción legible del error
          example: "Los datos de la solicitud no son válidos."
        timestamp:
          type: string
          format: date-time
          description: Instante (UTC, ISO-8601) en que se generó el error
          example: "2026-07-05T18:30:00Z"
        details:
          type: array
          description: >-
            Lista opcional de detalles/violaciones del error (cadenas). Se omite del
            cuerpo cuando no aplica.
          items:
            type: string
          example:
            - "La contraseña debe tener al menos 8 caracteres"
      examples:
        validacionFallida:
          summary: Error de validación
          value:
            error: "VALIDATION_ERROR"
            message: "Los datos de la solicitud no son válidos."
            timestamp: "2026-07-05T18:30:00Z"
        passwordInvalida:
          summary: Error con lista de detalles
          value:
            error: "INVALID_PASSWORD"
            message: "La contraseña no cumple los requisitos"
            timestamp: "2026-07-05T18:30:00Z"
            details:
              - "La contraseña debe tener al menos 8 caracteres"
...
    PaginatedResponseMeta:
      type: object
      description: Metadatos de paginación compartidos por todas las respuestas paginadas
      required:
        - total
        - page
        - pageSize
      properties:
        total:
          type: integer
          format: int64
          description: Total de elementos en la colección completa
          example: 47
        page:
          type: integer
          description: Índice de la página actual (base 0)
          example: 0
        pageSize:
          type: integer
          description: Número de elementos por página
          example: 20

    # --- Autenticación -------------------------------------------------------
...
    ParticipanteRequest:
      type: object
      description: |
        Participante adicional para una reserva. Exactamente uno de `userId` o `externalName`
        debe estar presente (no ambos, no ninguno).
      properties:
        userId:
          type: integer
          format: int64
          description: ID del usuario registrado en el sistema
          nullable: true
          example: 7
        externalName:
          type: string
          description: Nombre del participante externo sin cuenta en el sistema
          nullable: true
          maxLength: 100
          example: "Carlos García"
    ParticipanteResponse:
      type: object
      description: Información de un participante en la reserva
      required:
        - id
        - nombre
        - statusPago
        - isOwner
        - joinedAt
      properties:
        id:
          type: integer
          format: int64
          description: Identificador del registro de participante
          example: 3
        userId:
          type: integer
          format: int64
          description: ID del usuario registrado (null si es participante externo)
          nullable: true
          example: 7
        nombre:
          type: string
          description: Nombre del participante (registrado o externo)
          example: "Carlos García"
        statusPago:
          $ref: '#/components/schemas/PaymentStatus'
        isOwner:
          type: boolean
          description: Indica si este participante es el titular de la reserva
          example: false
        slotPosition:
          type: integer
          description: Posición en la reserva (1 = titular, 2-4 = resto)
          minimum: 1
          maximum: 4
          example: 2
        joinedAt:
          type: string
          format: date-time
          description: Instante en que se incorporó a la reserva
          example: "2025-06-14T18:30:00Z"
    ReservaRequest:
      type: object
      description: Datos para crear una nueva reserva
      required:
        - reservationDate
        - startTime
        - durationMinutes
      properties:
        reservationDate:
          type: string
          format: date
          description: Fecha de la reserva (no puede ser en el pasado)
          example: "2025-06-15"
        startTime:
          type: string
          description: Hora de inicio (en punto o en media hora, formato HH:mm)
          pattern: '^([01]\d|2[0-3]):[0-5]\d$'
          example: "09:00"
        durationMinutes:
          type: integer
          description: "Duración en minutos. Valores permitidos: 60, 90, 120, 150, 180"
          enum:
            - 60
            - 90
            - 120
            - 150
            - 180
          example: 60
        notes:
          type: string
          description: Observaciones libres del titular
          nullable: true
          maxLength: 500
          example: "Reserva para partido de entrenamiento"
        participantesAdicionales:
          type: array
          description: Participantes adicionales (máx. 3 — el creador ocupa el slot 1)
          items:
            $ref: '#/components/schemas/ParticipanteRequest'
          maxItems: 3
    ReservaResponse:
      type: object
      description: Datos completos de una reserva
      required:
        - id
        - reservationDate
        - startTime
        - durationMinutes
        - status
        - channel
        - priceTotal
        - participants
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único de la reserva (UUID)
          example: "550e8400-e29b-41d4-a716-446655440000"
        reservationDate:
          type: string
          format: date
          description: Fecha de la reserva
          example: "2025-06-15"
        startTime:
          type: string
          description: Hora de inicio (HH:mm)
          example: "09:00"
        endTime:
          type: string
          description: Hora de fin calculada (HH:mm)
          example: "10:00"
        durationMinutes:
          type: integer
          description: Duración en minutos
          example: 60
        status:
          $ref: '#/components/schemas/ReservationStatus'
        channel:
          $ref: '#/components/schemas/ReservationChannel'
        ownerId:
          type: integer
          format: int64
          description: ID del titular de la reserva
          example: 5
        ownerName:
          type: string
          description: Nombre completo del titular
          example: "John Doe"
        priceTotal:
          type: number
          format: double
          description: Importe total congelado en el momento de crear la reserva
          example: 15.00
        notes:
          type: string
          description: Observaciones libres del titular
          nullable: true
          example: "Reserva para partido de entrenamiento"
        participants:
          type: array
          description: Lista de participantes (incluye el titular como slot 1)
          items:
            $ref: '#/components/schemas/ParticipanteResponse'
        pago:
          description: |
            Pago asociado a la reserva (relación 1:1, creado atómicamente, RN-RES-03/D3).
            En el MVP (CASH) nace en estado `PENDING` con `amount` congelado.
          allOf:
            - $ref: '#/components/schemas/PagoResponse'
          nullable: true
        telegramMessageId:
          type: string
          description: ID del mensaje publicado en el grupo de Telegram (null si no aplica)
          nullable: true
          example: "12345"
        createdAt:
          type: string
          format: date-time
          description: Instante de creación de la reserva
          example: "2025-06-14T18:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Última modificación de la reserva
          example: "2025-06-14T18:00:00Z"
...
    UnirseResponse:
      type: object
      description: Confirmación de incorporación a una reserva
      required:
        - participanteId
        - reservaId
        - userId
        - nombre
        - statusPago
      properties:
        participanteId:
          type: integer
          format: int64
          description: ID del registro de participante creado
          example: 8
        reservaId:
          type: string
          format: uuid
          description: UUID de la reserva a la que se ha unido
          example: "550e8400-e29b-41d4-a716-446655440000"
        userId:
          type: integer
          format: int64
          description: ID del usuario que se ha unido
          example: 3
        nombre:
          type: string
          description: Nombre del usuario incorporado
          example: "Jane Smith"
        statusPago:
          $ref: '#/components/schemas/PaymentStatus'
...
    PaginatedReservasResponse:
      allOf:
        - $ref: '#/components/schemas/PaginatedResponseMeta'
        - type: object
          required:
            - data
          properties:
            data:
              type: array
              description: Lista de reservas de la página actual
              items:
                $ref: '#/components/schemas/ReservaResponse'

    # --- Pagos ---------------------------------------------------------------
    PagoResponse:
      type: object
      description: Datos de un pago asociado a una reserva
      required:
        - id
        - reservaId
        - amount
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          description: Identificador único del pago (UUID)
          example: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
        reservaId:
          type: string
          format: uuid
          description: UUID de la reserva asociada
          example: "550e8400-e29b-41d4-a716-446655440000"
        amount:
          type: number
          format: double
          description: Importe del pago en EUR
          example: 15.00
        method:
          $ref: '#/components/schemas/PaymentMethod'
        status:
          $ref: '#/components/schemas/PaymentStatus'
        redsysOrderId:
          type: string
          description: Referencia de orden Redsys (Ds_Merchant_Order)
          nullable: true
          example: "ORDER-20250615-001"
        paymentUrl:
          type: string
          format: uri
          description: URL del TPV virtual Redsys (solo disponible mientras el pago está en curso)
          nullable: true
          example: "https://sis-t.redsys.es:25443/sis/realizarPago"
        gateway:
          type: string
          description: Pasarela de pago utilizada
          nullable: true
          enum:
            - REDSYS
            - STRIPE
            - PAYPAL
          example: REDSYS
        transactionId:
          type: string
          description: Código de autorización del banco (Ds_AuthorisationCode)
          nullable: true
          example: "123456"
        paidAt:
          type: string
          format: date-time
          description: Instante de confirmación del pago
          nullable: true
          example: "2025-06-14T10:30:00Z"
        createdAt:
          type: string
          format: date-time
          description: Instante de creación del registro de pago
          example: "2025-06-14T09:00:00Z"
        updatedAt:
          type: string
          format: date-time
          description: Última modificación del pago
          example: "2025-06-14T10:30:00Z"
...